dom/reporting/tests/test_delivering_chips.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for CHIPS cookie in reporting endpoint response</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript">

// This test verifies that Firefox does not crash when the reporting endpoint
// responds with a Set-Cookie header that sets a CHIPS (Partitioned) cookie.

SpecialPowers.pushPrefEnv({ set: [
  ["dom.reporting.enabled", true],
  ["dom.reporting.header.enabled", true],
  ["dom.reporting.delivering.timeout", 1],
  ["dom.reporting.delivering.maxFailures", 2],
  ["dom.reporting.delivering.maxReports", 3],
]})

.then(_ => {
  window.addEventListener("message", e => {
    if (e.data.type == "finish") {
      ok(true, "No crash after reporting endpoint responded with CHIPS cookie");
      SimpleTest.finish();
      return;
    }

    if (e.data.type == "test") {
      ok(e.data.check, e.data.msg);
      return;
    }

    ok(false, "Invalid message");
  });

  let ifr = document.createElement("iframe");
  ifr.src = "https://example.org/tests/dom/reporting/tests/delivering.sjs?task=iframe&chips";

  document.body.appendChild(ifr);
});

SimpleTest.waitForExplicitFinish();

</script>
</body>
</html>
